second commit. Implemented jdbc for Admin, created DBUtils file and written logic for validation of user,fetch admin and create admin

+++ src/com/lt/bean/User.java
@@ -2,7 +2,7 @@ package com.lt.bean;
 
 public class User {
 
-	private long userId;
+	private int userId;
 	private String userName;
 	private String userPassword;
 	private int roleId;
@@ -16,11 +16,11 @@ public class User {
 	}
 
 	// role define. come from ENUM part role dynamic then Database, choice
-	public long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 
-	public void setUserId(long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
 
@@ -40,7 +40,7 @@ public class User {
 		this.userPassword = userPassword;
 	}
 
-	public User(long userId, String userName, String userPassword, int roleId) {
+	public User(int userId, String userName, String userPassword, int roleId) {
 		super();
 		this.userId = userId;
 		this.userName = userName;
